+++ sbuildr/generator/__init__.py
@@ -1,2 +1,0 @@
-from sbuildr.generator.rbuild import RBuildGenerator
-from sbuildr.generator.generator import Generator